justinScroll now works!
Proof that my idea for scrolling will work.

Uses self-modifying code.

Reduces byte footprint for number of animation frames for scrolling (to the right only for now) from 7 to 1 (plus a constant 37 bytes).

In theory, this will reduce frame count by a factor of 7, while only adding a constant byte size of ~160 (20 animation frames).  I believe this will be worth it in the long run.